added user authorization for CRUD routes

+++ main.py
@@ -49,11 +49,11 @@ def read_me(user_id: str = Depends(get_current_user_id)):
     return {"user_id": user_id}
 
 @app.get("/items/{item_id}", response_model=ItemOut)
-def read_item(item_id: int, conn = Depends(get_conn)):
+def read_item(item_id: int, conn = Depends(get_conn), user_id: str = Depends(get_current_user_id)):
     with conn.cursor() as cur:
         cur.execute(
-            "SELECT item_id, name, price, in_stock FROM items WHERE item_id = %s", 
-            (item_id,)
+            "SELECT item_id, name, price, in_stock FROM items WHERE item_id = %s AND user_id = %s", 
+            (item_id, user_id)
         )
         row = cur.fetchone()
     if row is None:
@@ -62,21 +62,21 @@ def read_item(item_id: int, conn = Depends(get_conn)):
     return to_output(row[0], saved)    
 
 @app.get("/items", response_model=list[ItemOut])
-def list_items(limit: int = 10, skip: int = 0, conn = Depends(get_conn)):
+def list_items(limit: int = 10, skip: int = 0, conn = Depends(get_conn), user_id: str = Depends(get_current_user_id)):
     with conn.cursor() as cur:
         cur.execute(
-            "SELECT item_id, name, price, in_stock FROM items ORDER BY item_id LIMIT %s OFFSET %s", (limit, skip)
+            "SELECT item_id, name, price, in_stock FROM items WHERE user_id=%s ORDER BY item_id LIMIT %s OFFSET %s", (user_id, limit, skip)
         )
         rows = cur.fetchall()
     return [to_output(item_id, Item(name=name, price=price, in_stock=in_stock)) for (item_id, name, price, in_stock) in rows]
 
 @app.post("/items", response_model=ItemOut, status_code=201)
-def create_item(item: Item, conn = Depends(get_conn)):
+def create_item(item: Item, conn = Depends(get_conn), user_id: str = Depends(get_current_user_id)):
     try:
         with conn.cursor() as cur:
             cur.execute(
-                "INSERT INTO items (name, price, in_stock) VALUES (%s, %s, %s) RETURNING item_id, name, price, in_stock",
-                (item.name, item.price, item.in_stock)
+                "INSERT INTO items (name, price, in_stock, user_id) VALUES (%s, %s, %s, %s) RETURNING item_id, name, price, in_stock, user_id",
+                (item.name, item.price, item.in_stock, user_id)
             )
             row = cur.fetchone()
 
@@ -86,30 +86,12 @@ def create_item(item: Item, conn = Depends(get_conn)):
     saved = Item(name=row[1], price=row[2], in_stock=row[3])
     return to_output(row[0], saved)
 
-@app.put("/items/{item_id}", response_model=ItemOut)
-def update_item(item_id: int, item: Item, conn = Depends(get_conn)):
-    try:
-        with conn.cursor() as cur:
-            cur.execute(
-                "UPDATE items SET name = %s, price = %s, in_stock =%s "
-                "WHERE item_id = %s "
-                "RETURNING item_id, name, price, in_stock",
-                (item.name, item.price, item.in_stock, item_id)
-            )
-            row = cur.fetchone()
-        if row is None:
-            raise HTTPException(status_code=404, detail="Item not found")
-    except psycopg.errors.UniqueViolation:
-        raise HTTPException(status_code=409, detail="Name already exists")
-    item_id, name, price, in_stock = row
-    return to_output(item_id, Item(name=name, price=price, in_stock=in_stock))
-
 @app.delete("/items/{item_id}", status_code=204)
-def delete_item(item_id: int, conn = Depends(get_conn)):
+def delete_item(item_id: int, conn = Depends(get_conn), user_id: str = Depends(get_current_user_id)):
     with conn.cursor() as cur:
         cur.execute(
-            "DELETE FROM items WHERE item_id = %s RETURNING item_id", 
-            (item_id,)
+            "DELETE FROM items WHERE item_id = %s AND user_id = %s RETURNING item_id", 
+            (item_id, user_id)
         )
         row = cur.fetchone()
 
@@ -118,11 +100,11 @@ def delete_item(item_id: int, conn = Depends(get_conn)):
     
 
 @app.patch("/items/{item_id}", response_model=ItemOut)
-def patch_item(item_id: int, update: ItemUpdate, conn = Depends(get_conn)):
+def patch_item(item_id: int, update: ItemUpdate, conn = Depends(get_conn), user_id: str = Depends(get_current_user_id)):
     with conn.cursor() as cur:
         cur.execute(
-            "SELECT item_id, name, price, in_stock FROM items WHERE item_id=%s",
-            (item_id,)
+            "SELECT item_id, name, price, in_stock FROM items WHERE item_id=%s AND user_id=%s",
+            (item_id, user_id)
         )
         row = cur.fetchone()
 
@@ -137,9 +119,9 @@ def patch_item(item_id: int, update: ItemUpdate, conn = Depends(get_conn)):
         with conn.cursor() as cur:
             cur.execute(
                 "UPDATE items SET name=%s, price=%s, in_stock=%s "
-                "WHERE item_id = %s "
+                "WHERE item_id = %s AND user_id = %s "
                 "RETURNING item_id, name, price, in_stock",
-                (update_data["name"], update_data["price"], update_data["in_stock"], item_id)
+                (update_data["name"], update_data["price"], update_data["in_stock"], item_id, user_id)
             )
             row = cur.fetchone()
